Fixed several problems regarding use of fixtures, naming and format
