+++ tests/integration/test_agentd/test_agentd_parametrized_reconnections/data/test_cases/cases_reconnection_protocol.yaml
@@ -15,7 +15,6 @@
     MAX_RETRIES: 5
     RETRY_INTERVAL: 5
     ENROLL: 'no'
-    DELETE_KEYS_FILE: true
   metadata:
 
 - name: Test_case_2
@@ -25,7 +24,6 @@
     MAX_RETRIES: 10
     RETRY_INTERVAL: 10
     ENROLL: 'no'
-    DELETE_KEYS_FILE: true
   metadata:
 
 - name: Test_case_3
@@ -35,7 +33,6 @@
     MAX_RETRIES: 5
     RETRY_INTERVAL: 5
     ENROLL: 'yes'
-    DELETE_KEYS_FILE: true
   metadata:
 
 - name: Test_case_4
@@ -45,7 +42,6 @@
     MAX_RETRIES: 2
     RETRY_INTERVAL: 2
     ENROLL: 'yes'
-    DELETE_KEYS_FILE: true
   metadata:
 
 - name: Test_case_5
@@ -55,7 +51,6 @@
     MAX_RETRIES: 1
     RETRY_INTERVAL: 1
     ENROLL: 'no'
-    DELETE_KEYS_FILE: true
   metadata:
 
 - name: Test_case_6
@@ -65,7 +60,6 @@
     MAX_RETRIES: 5
     RETRY_INTERVAL: 5
     ENROLL: 'no'
-    DELETE_KEYS_FILE: true
   metadata:
 
 - name: Test_case_7
@@ -75,7 +69,6 @@
     MAX_RETRIES: 10
     RETRY_INTERVAL: 4
     ENROLL: 'no'
-    DELETE_KEYS_FILE: true
   metadata:
 
 - name: Test_case_8
@@ -85,5 +78,4 @@
     MAX_RETRIES: 3
     RETRY_INTERVAL: 12
     ENROLL: 'no'
-    DELETE_KEYS_FILE: true
   metadata:
