+++ tests/integration/test_agentd/test_agentd_reconnection/test_agentd_reconection_enrollment_no_keys.py
@@ -58,7 +58,7 @@ from wazuh_testing.tools.simulators.authd_simulator import AuthdSimulator
 from wazuh_testing.utils.configuration import get_test_cases_data, load_configuration_template
 
 from . import CONFIGS_PATH, TEST_CASES_PATH
-from .. import wait_keepalive, wait_enrollment, kill_server, wait_enrollment_try
+from utils import wait_keepalive, wait_enrollment, wait_enrollment_try
 
 # Marks
 pytestmark = pytest.mark.tier(level=0)
@@ -143,7 +143,7 @@ def test_agentd_reconection_enrollment_no_keys(test_metadata, set_wazuh_configur
     wait_keepalive()    
 
     # Reset simulator
-    kill_server(remoted_server)
+    remoted_server.destroy()
 
     # Start rejecting Agent
     remoted_server = RemotedSimulator(protocol = test_metadata['PROTOCOL'], mode = 'WRONG_KEY')
@@ -153,7 +153,7 @@ def test_agentd_reconection_enrollment_no_keys(test_metadata, set_wazuh_configur
     wait_enrollment_try()
 
     # Reset simulator
-    kill_server(remoted_server)
+    remoted_server.destroy()
     
     # Start RemotedSimulator
     remoted_server = RemotedSimulator(protocol = test_metadata['PROTOCOL'])
@@ -163,7 +163,7 @@ def test_agentd_reconection_enrollment_no_keys(test_metadata, set_wazuh_configur
     wait_keepalive()    
 
     # Reset simulator
-    kill_server(authd_server)
+    authd_server.destroy()
 
     # Reset simulator
-    kill_server(remoted_server)
+    remoted_server.destroy()
