+++ tests/integration/test_agentd/test_agentd_reconnection/test_agentd_reconection_enrollment_with_keys.py
@@ -55,11 +55,10 @@ from wazuh_testing.constants.platforms import WINDOWS
 from wazuh_testing.modules.agentd.configuration import AGENTD_DEBUG, AGENTD_WINDOWS_DEBUG, AGENTD_TIMEOUT
 from wazuh_testing.tools.simulators.authd_simulator import AuthdSimulator
 from wazuh_testing.tools.simulators.remoted_simulator import RemotedSimulator
-from wazuh_testing.utils.client_keys import add_client_keys_entry
 from wazuh_testing.utils.configuration import get_test_cases_data, load_configuration_template
 
 from . import CONFIGS_PATH, TEST_CASES_PATH
-from .. import wait_keepalive, wait_enrollment_try, kill_server
+from utils import wait_keepalive, wait_enrollment_try
 
 # Marks
 pytestmark = pytest.mark.tier(level=0)
@@ -141,7 +140,7 @@ def test_agentd_reconection_enrollment_with_keys(test_configuration, test_metada
     wait_keepalive()
 
     # Reset simulator
-    kill_server(remoted_server)
+    remoted_server.destroy()
     
     # Start rejecting Agent
     remoted_server = RemotedSimulator(protocol = test_metadata['PROTOCOL'], mode = 'WRONG_KEY')
@@ -155,7 +154,7 @@ def test_agentd_reconection_enrollment_with_keys(test_configuration, test_metada
     wait_enrollment_try()
 
     # Reset simulator
-    kill_server(remoted_server)
+    remoted_server.destroy()
     
     # Start responding Agent
     remoted_server = RemotedSimulator(protocol = test_metadata['PROTOCOL'])
@@ -165,5 +164,5 @@ def test_agentd_reconection_enrollment_with_keys(test_configuration, test_metada
     wait_keepalive()
 
     # Reset simulator
-    kill_server(remoted_server)
-    kill_server(authd_server)
+    remoted_server.destroy()
+    authd_server.destroy()
